# Revert "epoll: atomically remove wait entry on wake up"